express_emitter/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
redis[hiredis]
loguru
prometheus-fastapi-instrumentator
httpx
psycopg2-binary
python-multipart
PyMuPDF
